+++ internal/run/upgrade.go
@@ -2,11 +2,8 @@ package run
 
 import (
 	"fmt"
-	"io/ioutil"
-	"os"
 
 	"github.com/pelotech/drone-helm3/internal/env"
-	"gopkg.in/yaml.v2"
 )
 
 // Upgrade is an execution step that calls `helm upgrade` when executed.
@@ -79,11 +76,15 @@ func (u *Upgrade) Prepare() error {
 
 	if u.chartVersion != "" {
 		args = append(args, "--version", u.chartVersion)
-		addChartVersion(u.chartVersion, u.chart)
+		EditChartYaml(u.chart, func(chartYaml *ChartYaml) {
+			chartYaml.Version = u.chartVersion
+		})
 	}
 
 	if u.appVersion != "" {
-		addAppVersion(u.appVersion, u.chart)
+		EditChartYaml(u.chart, func(chartYaml *ChartYaml) {
+			chartYaml.AppVersion = u.appVersion
+		})
 	}
 
 	if u.dryRun {
@@ -137,84 +138,4 @@ func (u *Upgrade) Prepare() error {
 	}
 
 	return nil
-}
-
-// ChartYaml is the struct representation of a Chart.yaml file.
-type ChartYaml struct {
-	APIVersion  string   `yaml:"apiVersion"`
-	Name        string   `yaml:"name"`
-	Version     string   `yaml:"version"`
-	Description string   `yaml:"description"`
-	AppVersion  string   `yaml:"appVersion,omitempty"`
-	Keywords    []string `yaml:"keywords,omitempty"`
-	Home        string   `yaml:"home,omitempty"`
-	Icon        string   `yaml:"icon,omitempty"`
-	Sources     []string `yaml:"sources,omitempty"`
-	Type        string   `yaml:"type,omitempty"`
-	Maintainers []struct {
-		Name  string `yaml:"name"`
-		Email string `yaml:"email"`
-		URL   string `yaml:"url,omitempty"`
-	} `yaml:"maintainers,omitempty"`
-	Dependencies []struct {
-		Name       string   `yaml:"name"`
-		Version    string   `yaml:"version"`
-		Repository string   `yaml:"repository"`
-		Condition  string   `yaml:"condition,omitempty"`
-		Tags       []string `yaml:"tags,omitempty"`
-	} `yaml:"dependencies,omitempty"`
-}
-
-func addAppVersion(appVersion, chartPath string) {
-
-	println("Updating Chart.yaml appVersion to " + appVersion)
-
-	yamlData, err := ioutil.ReadFile(chartPath + "/Chart.yaml")
-	if err != nil {
-		println("Try to read a file " + err.Error())
-	}
-	var chartYaml ChartYaml
-	if err := yaml.Unmarshal(yamlData, &chartYaml); err != nil {
-		println("Try to unmarshall the file: " + err.Error())
-	}
-
-	chartYaml.AppVersion = appVersion
-
-	updateYaml, err := yaml.Marshal(chartYaml)
-	if err != nil {
-		println("Try to marshall the file " + err.Error())
-	}
-
-	err = ioutil.WriteFile(chartPath+"/Chart.yaml", updateYaml, os.ModePerm)
-	if err != nil {
-		println("try to write the file" + err.Error())
-	}
-
-}
-
-func addChartVersion(chartVersion, chartPath string) {
-
-	println("Updating Chart.yaml version to " + chartVersion)
-
-	yamlData, err := ioutil.ReadFile(chartPath + "/Chart.yaml")
-	if err != nil {
-		println("Try to read a file " + err.Error())
-	}
-	var chartYaml ChartYaml
-	if err := yaml.Unmarshal(yamlData, &chartYaml); err != nil {
-		println("Try to unmarshall the file: " + err.Error())
-	}
-
-	chartYaml.Version = chartVersion
-
-	updateYaml, err := yaml.Marshal(chartYaml)
-	if err != nil {
-		println("Try to marshall the file " + err.Error())
-	}
-
-	err = ioutil.WriteFile(chartPath+"/Chart.yaml", updateYaml, os.ModePerm)
-	if err != nil {
-		println("try to write the file" + err.Error())
-	}
-
 }
